Refactor: erradicação de magic strings e desmembramento de God Objects
- **Sistema de Tipagem Forte (Racas/Classes):**
  - Implementados enums `TipoRaca` e `TipoClasse` para substituição de strings mágicas hardcoded.
  - Criadas `FabricaRacas` e `FabricaClasses` para instanciar personagens dinamicamente via Enum.
  - Atualizados os menus de criação de personagem e a lógica do `isBoss()` para depender inteiramente da tipagem por Enum.
- **Desmembramento do Combate.cpp:**
  - Extração de regras de negócio de batalha para a pasta `src/Sistemas/Combate/Mecanicas/`.
  - Criada `CalculadoraDano` (isolando a matemática de hits, críticos e evasão).
  - Criado `GerenciadorTurnos` (isolando a iniciativa e fluxo de combate).
  - Criada `MecanicasInimigo` (isolando comportamento e ações padrão do inimigo).
- **Desmembramento do ControleMapa.cpp:**
  - Extração da exploração do mapa para múltiplos especialistas na pasta `src/Mapas/Sistemas/`.
  - Criada `FisicaMapa` para verificação de barreiras e limites.
  - Criado `CarregadorMapa` para leitura e instanciamento de matrizes e tiles especiais.
  - Criado `RenderizadorMapa` para lidar com a exibição do minimapa.
  - Criado `ControladorInputMapa` focado apenas no processamento de teclado/menus.
  - Criado `AnimadorMapa` para orquestrar as artes ASCII de introdução.
  - Redução massiva de linhas no `ControleMapa.cpp`, atuando agora apenas como um coordenador central de delegados.
- **Documentação:**
  - Adicionado `Relatorio_Refatoracao.md` descrevendo a nova arquitetura do projeto e o cumprimento das metodologias SOLID implementadas hoje.

+++ src/Sistemas/ConfiguracaoTerminal/Aparencia/Aparencia.h
@@ -70,7 +70,7 @@ public:
     static int lerInteiroEmPopupFlutuante(const std::string& mensagem, int limiteMin, int limiteMax, Cor corTema = Cor::BRANCO);
     static void limparPopupAnterior();
     static void definirUltimoPopup(int x, int y, int w, int h, const std::string& bgAnsi = "\033[48;2;25;25;25m");
-    static std::vector<std::string> reduzirEscalaAscii(const std::vector<std::string>& arteOriginal, int fatorX = 2, int fatorY = 2);
+    static std::vector<std::string> reduzirEscalaAscii(const std::vector<std::string>& arteOriginal, float fatorX = 2.0f, float fatorY = 2.0f);
     static std::string sobreporLogoAnsi(const std::string& backgroundLine, const std::vector<std::string>& logoChars, int startX, const std::string& fgColor, int larguraTerminal);
     static std::string fadarLinhaAnsi(const std::string& linha, float ratio);
